refactor(ledger): drop Huma suffixes on operation route :hammer:
The five exported fiber.Ctx terminals on OperationRouteHandler were
unreachable — registerOperationRouteRoutesToApp passes middleware only and
hands the terminals to Huma. As with instrument, they were interleaved with the
cores rather than gathered in one trailing block, so each was removed
individually and a cores banner added to match the converted siblings.

- Delete the dead Fiber wrappers and operation_route_test.go, the only one of
  the six operation_route test files that mounted them. The validation,
  block/unblock, merge-overdraft, overdraft-refund and v2-register files test
  validators and the OpenAPI document directly and are kept untouched.
- Port the branches operation_route_test.go covered onto the Huma transport;
  every core holds its baseline. Deleting it without the ports would have
  dropped createOperationRoute to 54.2%, getOperationRouteByID to 66.7%,
  updateOperationRoute to 67.6%, deleteOperationRouteByID to 62.5%,
  getAllOperationRoutes to 60.7% and validateAccountRule to 18.2%.
- Rename methods and envelope types: CreateOperationRouteHuma ->
  CreateOperationRoute, and the nine envelope types to Request/Response.
- operation_route_handler_huma.go -> operation_route_handler.go,
  operation_route_huma_test.go -> operation_route_handler_test.go; fix the
  filename citation in transaction_route_handler_huma.go.
- permissions.yaml: the operation-routes citation pointed at routes.go:315,
  which is the transaction-routes registrar — a pre-existing error. It now
  names routes.go:273, the first routePost of its own registrar.

No route, operation ID, or wire contract changes: check-docs reports no
drift.

X-Lerian-Ref: 0x1
